Move smina results processing into plugin
